Redid DriveERRFix -> PermFoldERR
Reworked this part, the user now has two options, if they changed the drive the application is launched from they'll be asked for the original drive name and it should be fixed. If NeaPermFolder is just gone they'll be run through the first time setup again, no data will be overwritten.